+++ pilot/pkg/dataplane/networking/deprecated/fault.go
