+++ pilot/pkg/dataplane/networking/deprecated/header.go
