+++ pilot/pkg/dataplane/networking/deprecated/util.go
